delete all the mouse code from last year
starting from scratch will help us reduce code duplication

+++ real/commands/LEDBlink.cpp
@@ -1,6 +1,6 @@
 #include "LEDBlink.h"
 
-LEDBlink::LEDBlink(const int led_pin, unsigned long blink_time) : Command("blink"), led_pin(led_pin),
+LEDBlink::LEDBlink(const uint8_t led_pin, unsigned long blink_time) : Command("blink"), led_pin(led_pin),
                                                                   blink_time(blink_time) {}
 
 void LEDBlink::initialize() {
